Replace spoken language flags with neutral badges
